// frontend/jobapptrackerfrontend/src/enums/jobinterviewtestids/EditInterviewModalTestIds_enum.ts
export enum EditInterviewModalTestIds {
    interviewType = "editInterviewTypeField",
    interviewTypeError = "editInterviewTypeError",
    locationField = "editInterviewLocationField",
    startDateField = "editInterviewStartDateField",
    startDateError = "editInterviewStartDateError",
    endDateField = "editInterviewEndDateField",
    endDateError = "editInterviewEndDateError",
    cancelButton = "editInterviewCancelButton",
    editInterviewButton = "editIntervieweditInterviewButton"
}